changed connection pooling to be custom and refactored packaging
